pyproject.toml
[tool.poetry]
name = "filestash-minio-oidc-auth"
version = "0.1.0"
description = ""
authors = ["NextFire <git@yuru.moe>"]
readme = "README.md"
packages = [{include = "filestash_minio_oidc_auth"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
aiohttp = "^3.8.4"
xmltodict = "^0.13.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
